feat(folderStructure)!: split repository into packages for a new project structure
