+++ config.go
@@ -6,8 +6,7 @@ import (
 )
 
 type Config struct {
-	Port  int    `envconfig:"PORT" default:"3000"`
-	Token string `envconfig:"TOKEN"`
+	Port int `envconfig:"PORT" default:"3000"`
 }
 
 func newConfig() (*Config, error) {
