+++ internalHandlers/pongHandler.go
@@ -1,0 +1,20 @@
+package internalHandlers
+
+import (
+	"encoding/json"
+	"net/http"
+
+	"github.com/Scalingo/go-utils/logger"
+)
+
+func PongHandler(w http.ResponseWriter, r *http.Request, _ map[string]string) error {
+	log := logger.Get(r.Context())
+	w.Header().Add("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+
+	err := json.NewEncoder(w).Encode(map[string]string{"status": "pong"})
+	if err != nil {
+		log.WithError(err).Error("Fail to encode JSON")
+	}
+	return nil
+}
